update arg checking; correct file deletion retaining final vcf; safeguard finds so we don't send in wildcard only

+++ runRufus.sh
@@ -451,11 +451,13 @@ clean_up_files ()
     fi
 
     # Remove files from sub directories for this region only
-	find ./Intermediates -maxdepth 1 -type f -name "*${formatted_region}*" -delete
-	find ./TempOverlap -maxdepth 1 -type f -name "*${formatted_region}*" -delete
-	find . -maxdepth 1 -type f -name "*${formatted_region}*generator*" -delete
-	find . -maxdepth 1 -type f -name "*${formatted_region}.txt" -delete
-	find . -maxdepth 1 -type p -name "*${formatted_region}*" -delete # Clean up pipes too
+	if [ "$formatted_region" != "" ]; then
+		find ./Intermediates -maxdepth 1 -type f -name "*${formatted_region}*" -delete
+		find ./TempOverlap -maxdepth 1 -type f -name "*${formatted_region}*" -delete
+		find . -maxdepth 1 -type f -name "*${formatted_region}*generator*" -delete
+		find . -maxdepth 1 -type f -name "*${formatted_region}.txt" -delete
+		find . -maxdepth 1 -type p -name "*${formatted_region}*" -delete # Clean up pipes too
+	fi
   else
     echo "not cleaning up files"
   fi
